Cleans up venues view page
On submitting new venue, styles ‘Venue Amenities’ so it is less
cluttered and easier to read. Also fixes database error in relation to
Amenities when trying to add new venue and when on venue view page.
Adds sql fields for amenities to sc_Venue since that is how the rest of
the app was coded to work. Can adjust based on other DB tables for
amenities if needed.